+++ package.json
@@ -17,7 +17,7 @@
     "db:setup": "node src/scripts/createCollections.js",
     "db:delete": "node src/scripts/deleteCollections.js",
     "db:reset": "npm run db:delete && npm run db:setup",
-    "connectathon-upload": "node src/scripts/getConnectathonBundles.js",
+    "upload-bundles": "node src/scripts/uploadPremadeBundles.js",
     "parse-compartment-def": "node src/scripts/parseCompartmentDefinition.js"
   },
   "repository": {
